:sparkles: Set up CI with Azure Pipelines on windows and macos, resolve #122. :wheel_chair: skip the file permission related tests for now. #141

+++ tests/test_hash_store.py
@@ -1,6 +1,8 @@
 import os
+import sys
 
 from moban.hashstore import HashStore
+from nose import SkipTest
 
 
 class TestHashStore:
@@ -13,7 +15,8 @@ class TestHashStore:
         )
 
     def tearDown(self):
-        os.unlink(".moban.hashes")
+        if os.path.exists(".moban.hashes"):
+            os.unlink(".moban.hashes")
 
     def test_simple_use_case(self):
         hs = HashStore()
@@ -76,6 +79,8 @@ class TestHashStore:
         Save as above, but this time,
         the generated file had file permision change
         """
+        if sys.platform == 'win32':
+            raise SkipTest('No actual chmod on windows')
         hs = HashStore()
         flag = hs.is_file_changed(*self.fixture)
         if flag:
